extracting AssociationFilter into an utility class to be used both in the helper and the task
